+++ NEWS.md
@@ -12,6 +12,7 @@ Version: development
 * The statistics can now return dataframe that is also the default (allows nicer printing)
 * The add_column can now take a position argument
 * The init with a table can now also accept column_order argument
+* Added has_batch and get_batch
 
 Version: 1.1
 -----------